requirements.txt
pillow
flask
requests
pyuwsgi
